LPS-90766 1.) Move to the same package as com.liferay.portal.kernel.servlet.DummyFilterChain 2.) match StubHttpServletResponse
